refactor(context): extract output helpers
